+++ cypress.json
@@ -1,5 +1,8 @@
 {
   "video": false,
-  "screenshotOnRunFailure": false
-
+  "screenshotOnRunFailure": false,
+  "testFiles": "**/*.cypress.*",
+  "integrationFolder": "tests/apps",
+  "fixturesFolder": "tests/utils/cypress/fixtures",
+  "pluginsFile": "tests/utils/cypress/plugins/index.ts"
 }
