--- cypress/start.ts
@@ -1,38 +1,0 @@
-import app from '../src/apps/backoffice/frontend/app';
-import cypress from 'cypress';
-import { Server } from 'http';
-import cypressConfig from '../cypress.json';
-
-async function run() {
-  const server = await startServer();
-  await runCypress();
-  server.close(() => {
-    process.exit(0);
-  });
-}
-
-async function startServer(): Promise<Server> {
-  let server: Server;
-
-  return new Promise((resolve, reject) => {
-    server = app.listen(app.get('port'), async () => {
-      console.log(`  Backoffice frontend is running at http://localhost:${app.get('port')} in ${app.get('env')} mode`);
-      console.log('  Press CTRL-C to stop\n');
-      resolve(server);
-    });
-  });
-}
-
-async function runCypress() {
-  return cypress.run({
-    reporter: 'junit',
-    browser: 'chrome',
-    headless: true,
-    config: {
-      ...cypressConfig,
-      baseUrl: `http://localhost:${app.get('port')}`,
-    }
-  });
-}
-
-run();
