+++ package.json
@@ -13,7 +13,7 @@
   "scripts": {
     "dev": "NODE_ENV=dev ts-node-dev --ignore-watch node_modules --inspect=0.0.0.0:9267 ./src/apps/mooc_backend/server.ts",
     "dev:backoffice:frontend": "NODE_ENV=dev ts-node-dev --ignore-watch node_modules ./src/apps/backoffice/frontend/server.ts",
-    "test": "npm run test:unit && npm run test:features",
+    "test": "npm run test:unit && npm run test:features && npm run cypress:run",
     "test:unit": "NODE_ENV=test jest",
     "test:features": "NODE_ENV=test cucumber-js -p default",
     "lint": "tslint src/**/*.ts{,x}",
@@ -23,8 +23,8 @@
     "build:tsc": "tsc -p tsconfig.prod.json",
     "build:di": "copy 'src/**/*.{json,yaml,html,png}' dist/src",
     "build:clean": "rm -r dist; exit 0",
-    "cypress:open": "NODE_ENV=test cypress open",
-    "cypress:run": "NODE_ENV=test ts-node cypress/start"
+    "cypress:open": "NODE_ENV=test ts-node tests/utils/cypress/open",
+    "cypress:run": "NODE_ENV=test ts-node tests/utils/cypress/run"
   },
   "dependencies": {
     "@types/bson": "^4.0.2",
