+++ tests/utils/cypress/open.ts
@@ -1,0 +1,16 @@
+import cypress from 'cypress';
+import { startBackofficeFrontend } from './startBackofficeFrontend';
+
+async function open() {
+  const server = await startBackofficeFrontend();
+  await openCypress();
+  server.close(() => {
+    process.exit(0);
+  });
+}
+
+async function openCypress() {
+  return cypress.open();
+}
+
+open();
