+++ tests/utils/cypress/tsconfig.json
@@ -1,11 +1,9 @@
 {
-  "extends": "../tsconfig.json",
+  "extends": "../../../tsconfig.json",
   "compilerOptions": {
     "lib": ["es2015", "dom"],
     "types": ["cypress"]
   },
-  "include": [
-    "**/*.ts"
-  ]
+  "include": ["**/*.ts"]
 }
